+++ src/commands/commands.ts
@@ -1,9 +1,10 @@
 import { commands, Disposable, ExtensionContext, Uri, window } from 'vscode';
-import { telemetry } from 'extension';
+
 import { showOutputChannel } from 'cli/shell/output';
+import { telemetry } from 'extension';
+import { CommandId } from 'types/extensionIds';
 import { TelemetryErrorEventNames } from 'types/telemetryEventNames';
 import { refreshAllTreeViews, refreshResourcesTreeViews } from 'ui/treeviews/treeViews';
-import { CommandId } from 'types/extensionIds';
 import { addKustomization } from './addKustomization';
 import { addSource } from './addSource';
 import { copyResourceName } from './copyResourceName';
