+++ src/commands/copyResourceName.ts
@@ -1,4 +1,5 @@
 import { env } from 'vscode';
+
 import { SourceNode } from 'ui/treeviews/nodes/sourceNode';
 import { WorkloadNode } from 'ui/treeviews/nodes/workloadNode';
