+++ src/commands/createSource.ts
@@ -1,7 +1,7 @@
 import gitUrlParse from 'git-url-parse';
 import { commands, env, Uri, window } from 'vscode';
-import { CommandId } from 'types/extensionIds';
 
+import { CommandId } from 'types/extensionIds';
 
 /**
  * Show notifications reminding users to add a public key
