+++ src/commands/fluxCheck.ts
@@ -1,4 +1,5 @@
 import safesh from 'shell-escape-tag';
+
 import { shell } from 'cli/shell/exec';
 import { ClusterContextNode } from 'ui/treeviews/nodes/clusterContextNode';
