+++ src/commands/fluxReconcileSource.ts
@@ -1,11 +1,12 @@
 import { window } from 'vscode';
+
 import { fluxTools } from 'cli/flux/fluxTools';
 import { FluxSource } from 'types/fluxCliTypes';
 import { KubernetesObjectKinds } from 'types/kubernetes/kubernetesTypes';
 import { BucketNode } from 'ui/treeviews/nodes/bucketNode';
 import { GitRepositoryNode } from 'ui/treeviews/nodes/gitRepositoryNode';
-import { OCIRepositoryNode } from 'ui/treeviews/nodes/ociRepositoryNode';
 import { HelmRepositoryNode } from 'ui/treeviews/nodes/helmRepositoryNode';
+import { OCIRepositoryNode } from 'ui/treeviews/nodes/ociRepositoryNode';
 import { refreshSourcesTreeView } from 'ui/treeviews/treeViews';
 
 /**
