+++ src/commands/openResource.ts
@@ -1,4 +1,5 @@
 import { Uri, window, workspace } from 'vscode';
+
 import { telemetry } from 'extension';
 import { TelemetryErrorEventNames } from 'types/telemetryEventNames';
