+++ src/commands/setClusterProvider.ts
@@ -1,6 +1,7 @@
 import { window } from 'vscode';
-import { globalState } from 'extension';
+
 import { ClusterMetadata } from 'data/globalState';
+import { globalState } from 'extension';
 import { KnownClusterProviders, knownClusterProviders } from 'types/kubernetes/kubernetesTypes';
 import { ClusterContextNode } from 'ui/treeviews/nodes/clusterContextNode';
 import { refreshAllTreeViews } from 'ui/treeviews/treeViews';
