+++ src/data/globalState.ts
@@ -1,4 +1,5 @@
 import { ExtensionContext, window, workspace } from 'vscode';
+
 import { KnownClusterProviders } from 'types/kubernetes/kubernetesTypes';
 
 export interface ClusterMetadata {
