+++ src/types/nodeContext.ts
@@ -2,7 +2,6 @@
  * Defines GitOps tree view node context values.
  */
 export const enum NodeContext {
-
 	// Cluster context values
 	Cluster = 'cluster',
 	CurrentCluster = 'currentCluster',
