+++ src/ui/treeviews/nodes/anyResourceNode.ts
@@ -1,4 +1,5 @@
 import { KubernetesObject } from '@kubernetes/client-node';
+
 import { TreeNode } from './treeNode';
 
 /**
