+++ src/ui/treeviews/nodes/clusterContextNode.ts
@@ -1,12 +1,13 @@
 import { ExtensionMode, MarkdownString } from 'vscode';
-import { extensionContext, globalState, setVSCodeContext } from 'extension';
+
 import { fluxVersion } from 'cli/checkVersions';
 import { currentContextName, kubernetesTools } from 'cli/kubernetes/kubernetesTools';
+import { extensionContext, globalState, setVSCodeContext } from 'extension';
 import { CommandId, ContextId } from 'types/extensionIds';
 import { KubernetesCluster, KubernetesContextWithCluster } from 'types/kubernetes/kubernetesConfig';
 import { ClusterProvider } from 'types/kubernetes/kubernetesTypes';
 import { createMarkdownHr, createMarkdownTable } from 'utils/markdownUtils';
-import { NodeContext } from './nodeContext';
+import { NodeContext } from '../../../types/nodeContext';
 import { TreeNode } from './treeNode';
 
 /**
