+++ src/ui/treeviews/nodes/gitRepositoryNode.ts
@@ -1,6 +1,6 @@
 import { GitRepository } from 'types/flux/gitRepository';
 import { KubernetesObjectKinds } from 'types/kubernetes/kubernetesTypes';
-import { NodeContext } from './nodeContext';
+import { NodeContext } from '../../../types/nodeContext';
 import { SourceNode } from './sourceNode';
 
 /**
