+++ src/ui/treeviews/nodes/treeNode.ts
@@ -1,11 +1,12 @@
 import { KubernetesObject } from '@kubernetes/client-node';
 import { Command, MarkdownString, ThemeColor, ThemeIcon, TreeItem, TreeItemCollapsibleState, Uri } from 'vscode';
-import { asAbsolutePath } from 'utils/asAbsolutePath';
+
 import { kubernetesTools } from 'cli/kubernetes/kubernetesTools';
 import { CommandId } from 'types/extensionIds';
 import { FileTypes } from 'types/fileTypes';
 import { GitOpsTemplate } from 'types/flux/gitOpsTemplate';
 import { KubernetesContextWithCluster } from 'types/kubernetes/kubernetesConfig';
+import { asAbsolutePath } from 'utils/asAbsolutePath';
 import { KnownTreeNodeResources, createMarkdownTable } from 'utils/markdownUtils';
 
 export const enum TreeNodeIcon {
