+++ src/ui/webviews/configureGitOps/receiveMessage.ts
@@ -1,4 +1,5 @@
 import { WebviewPanel } from 'vscode';
+
 import { actionCreate, actionYAML } from './actions';
 
 export async function receiveMessage(message: any, panel: WebviewPanel) {
