+++ src/utils/asAbsolutePath.ts
@@ -1,4 +1,5 @@
 import { Uri } from 'vscode';
+
 import { extensionContext } from 'extension';
 
 export function asAbsolutePath(relativePath: string): Uri {
